feat(nextls): multi-root workspaces
This patch has a dependency on the `mhanberg/workspace-folders.nvim`
plugin.

To try out the behavior, intsall that plugin and create a
`.code-workspace` file that describes your multi-root workspace (or mono
repo as most call it).

If your monorepo is called "money-factory-io", with the folders
"crypto", "ai-chat-app", and "drop-shipping-cms", your `code-workspace`
would be named `money-factory-io.code-workspace` and look like:

```json
{
  "folders": [
  {"path": "crypto"},
  {"path": "ai-chat-app"},
  {"path": "drop-shipping-cms"},
  ]
}
```